requirements.txt
Flask>=2.0.0,<3.1.0
mysql-connector-python>=8.0.0,<8.1.0
Werkzeug>=2.0.0,<3.1.0
Jinja2>=3.0.0,<3.2.0
itsdangerous>=2.0.0,<2.2.0
click>=8.0.0,<8.2.0